+++ src/iutransform/reduce.cpp
@@ -63,8 +63,8 @@ IuStatus reduce(const iu::ImageGpu_32f_C1* src, iu::ImageGpu_32f_C1* dst,
     float x_factor = (float)dst->width() / (float)src->width();
     float y_factor = (float)dst->height() / (float)src->height();
 
-    float sigma =/*0.5774f*/0.3f * sqrt(0.5f*(x_factor+y_factor));
-    //float sigma = 0.1f * sqrt(0.5f*(x_factor+y_factor));
+    float sigma =/*0.5774f*/0.3f * sqrtf(0.5f*(x_factor+y_factor));
+    //float sigma = 0.1f * sqrtf(0.5f*(x_factor+y_factor));
     unsigned int kernel_size = 5;
 
     iuprivate::filterGauss(src, filtered, src->roi(), sigma, kernel_size);
